Fix code smells found by Sonar Cloud (#831)
* Fix bugs found by SonarCloud

* Fix bugs and major code smells found by SonarCloud

* Fix critical code smells found by SonarCloud

* Fix major code smells found by sonar cloud

* Fix critical code smells found by SonarCloud

* Fix critical code smells found by SonarCloud

* Fix critical code smells found by sonar cloud

* Fix major code smells found by sonar cloud

* Fix major code smells in ethernet.c

* Fix major code smells in fw_update.c

* Fix major code smells

* Fix major code smells in http.c

* Fix major code smell found by sonar cloud

* Refactoring

* Fix major code smells

* Refactoring

* Refactoring

* Refactoring

* Refactoring

* Refactoring

* Fix minor code smells

* Refactoring to reduce cognitive complexity and fix critical code smells

* Refactoring

* Refactoring, update ruuvi.json_stream_gen.c

* Disable unit-tests for test_leds and test_time_task

* Refactoring: split function to fix major code smell

* Refactoring: extract gw_cfg_cmp from gw_cfg.c

* Refactoring: optimize size of gw_cfg_json_parse.c

* Add NOSONAR comments

* Split http.c to several files

* Extract validate_url.c from http_server_cb_on_get.c

* Small refactoring, add logs

* Update esp32-wifi-manager and ruuvi.json_stream_gen.c submodules to master branches

* hmac_sha256: fix bug: reset json_stream_gen after HMAC generation

* Remove deprecated json_helper.c

* Run sonar_scan in separate folder 'build-sonar'

* Refactoring http_server_cb_on_get: firmware updates

* Update esp32-wifi-manager to the master branch

+++ main/nrf52fw.c
@@ -213,7 +213,7 @@ nrf52fw_parse_segment_info_line(const char* p_version_line, nrf52fw_segment_t* p
         end += 1;
     }
     const char*  p_segment_file_name_end = end;
-    const size_t segment_file_name_len   = (size_t)(ptrdiff_t)(p_segment_file_name_end - p_segment_file_name_begin);
+    const size_t segment_file_name_len   = (size_t)(p_segment_file_name_end - p_segment_file_name_begin);
     if (segment_file_name_len >= sizeof(p_segment->file_name))
     {
         return false;
